Player AttackLevelUpUI
Player Attack, LevelUpUI만드는중

+++ Assets/01.Unit/UnitData.cs
@@ -46,7 +46,7 @@ public struct UnitStat
 {
     public int maxHp;
     public float moveSpeed;
-    public float defensePower;
+    public int defensePower;
     public float attackRange;
     public int attackPower;
 }
